backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pytesseract==0.3.10
Pillow==10.2.0
reportlab==4.0.9
google-auth==2.27.0
requests==2.31.0
python-dotenv==1.0.1
